+++ mjx/mujoco/mjx/_src/types.py
@@ -771,6 +771,7 @@ class Model(PyTreeNode):
     tendon_margin: min distance for limit detection           (ntendon,)
     tendon_stiffness: stiffness coefficient                   (ntendon,)
     tendon_damping: damping coefficient                       (ntendon,)
+    tendon_armature: inertia associated with tendon velocity  (ntendon,)
     tendon_frictionloss: loss due to friction                 (ntendon,)
     tendon_lengthspring: spring resting length range          (ntendon, 2)
     tendon_length0: tendon length in qpos0                    (ntendon,)
@@ -844,6 +845,7 @@ class Model(PyTreeNode):
     name_tupleadr: tuple name pointers                        (ntuple,)
     name_keyadr: keyframe name pointers                       (nkey,)
     names: names of all objects, 0-terminated                 (nnames,)
+    signature: compilation signature
   """
 
   nq: int
@@ -1112,6 +1114,7 @@ class Model(PyTreeNode):
   tendon_margin: jax.Array
   tendon_stiffness: jax.Array
   tendon_damping: jax.Array
+  tendon_armature: jax.Array
   tendon_frictionloss: jax.Array
   tendon_lengthspring: jax.Array
   tendon_length0: jax.Array
@@ -1187,6 +1190,7 @@ class Model(PyTreeNode):
   name_tupleadr: np.ndarray
   name_keyadr: np.ndarray
   names: bytes
+  signature: np.uint64
   _sizes: jax.Array