+++ unity/Runtime/Bindings/MjBindings.cs
@@ -320,6 +320,7 @@ public enum mjtObj : int{
   mjNOBJECT = 26,
   mjOBJ_FRAME = 100,
   mjOBJ_DEFAULT = 101,
+  mjOBJ_MODEL = 102,
 }
 public enum mjtConstraint : int{
   mjCNSTR_EQUALITY = 0,
@@ -5577,6 +5578,7 @@ public unsafe struct mjModel_ {
   public double* tendon_margin;
   public double* tendon_stiffness;
   public double* tendon_damping;
+  public double* tendon_armature;
   public double* tendon_frictionloss;
   public double* tendon_lengthspring;
   public double* tendon_length0;
@@ -5764,6 +5766,7 @@ public unsafe struct mjsCompiler_ {
   public byte fusestatic;
   public int inertiafromgeom;
   public fixed int inertiagrouprange[2];
+  public byte saveinertial;
   public int alignfree;
   public mjLROpt_ LRopt;
 }