+++ doc/changelog.rst
@@ -10,10 +10,15 @@ Upcoming version (not yet released)
 
    - The default value of the flag for toggling :ref:`internal flex contacts<flex-contact-internal>` was changed from
      "true" to "false". This feature has proven to be counterintuitive for users.
+   - All of the attach functions (``mjs_attachBody``, ``mjs_attachFrame``, ``mjs_attachToSite``,
+     ``mjs_attachFrameToSite``) have been removed and replaced by a single function :ref:`mjs_attach`.
 
 General
 ^^^^^^^
-- Add :ref:`orientation<body-composite-quat>` parameter to :ref:`composite<body-composite>`. Moreover, allow the
+- Added :ref:`tendon armature<tendon-spatial-armature>`: inertia associated with changes in tendon length.
+- Added the :ref:`compiler/saveinertial<compiler-saveinertial>` flag, writing explicit inertial clauses for all
+  bodies when saving to XML.
+- Added :ref:`orientation<body-composite-quat>` attribute to :ref:`composite<body-composite>`. Moreover, allow the
   composite to be the direct child of a frame.
 
 Bug fixes
@@ -22,10 +27,16 @@ Bug fixes
   which the Jacobian is computed, now fixed.
 - Fixed a bug that caused the parent frame of elements in the child worldbody to be incorrectly set when attaching an
   mjSpec to a frame or a site.
+- Fixed a bug that caused shadow rendering to flicker on platforms (e.g., MacOS) that do not support ARB_clip_control.
+  Fixed in collaboration with :github:user:`aftersomemath`.
+
+Python bindings
+^^^^^^^^^^^^^^^
+- Added support for nameless :ref:`mjSpec` objects in the ``bind`` method, see the corresponding :ref:`section<PyMJCF>`
+  in the documentation.
 
 Version 3.3.0 (Feb 26, 2025)
 ----------------------------
-
 
 Feature promotion
 ^^^^^^^^^^^^^^^^^